+++ Additionals/ChangeLog.md
@@ -3,6 +3,7 @@
 ## Version 1.3.1
 
 * Recognition service added for automatic detection of sample values
+* Request is shown if a sample with value shall be removed
 * Menu entries reordered
 
 ## Version 1.3.0